+++ tests/providers/Data.php
@@ -1,0 +1,144 @@
+<?php
+
+namespace yiiunit\extensions\redis\providers;
+
+final class Data
+{
+    public static function acquireTimeout()
+    {
+        return [
+            'no timeout (lock is held)' => [0, false, false],
+            '2s (lock is held)' => [1, false, false],
+            '3s (lock will be auto released in acquire())' => [2, true, false],
+            '3s (lock is auto released)' => [2, true, true],
+        ];
+    }
+
+    public static function hmSetData(): array
+    {
+        return [
+            [
+                ['hmset1', 'one', '1', 'two', '2', 'three', '3'],
+                [
+                    'one' => '1',
+                    'two' => '2',
+                    'three' => '3'
+                ],
+            ],
+            [
+                ['hmset2', 'one', null, 'two', '2', 'three', '3'],
+                [
+                    'one' => '',
+                    'two' => '2',
+                    'three' => '3'
+                ],
+            ]
+        ];
+    }
+
+    public static function illegalValuesForFindByCondition(): array
+    {
+        return [
+            // code injection
+            [['id' => ["' .. redis.call('FLUSHALL') .. '" => 1]], ["'\\' .. redis.call(\\'FLUSHALL\\') .. \\'", 'rediscallFLUSHALL'], ["' .. redis.call('FLUSHALL') .. '"]],
+            [['id' => ['`id`=`id` and 1' => 1]], ["'`id`=`id` and 1'", 'ididand']],
+            [['id' => [
+                'legal' => 1,
+                '`id`=`id` and 1' => 1,
+            ]], ["'`id`=`id` and 1'", 'ididand']],
+            [['id' => [
+                'nested_illegal' => [
+                    'false or 1=' => 1
+                ]
+            ]], [], ['false or 1=']],
+
+            // custom condition injection
+            [['id' => [
+                'or',
+                '1=1',
+                'id' => 'id',
+            ]], ["cid0=='or' or cid0=='1=1' or cid0=='id'"], []],
+            [['id' => [
+                0 => 'or',
+                'first' => '1=1',
+                'second' => 1,
+            ]], ["cid0=='or' or cid0=='1=1' or cid0=='1'"], []],
+            [['id' => [
+                'name' => 'test',
+                'email' => 'test@example.com',
+                "' .. redis.call('FLUSHALL') .. '" => "' .. redis.call('FLUSHALL') .. '"
+            ]], ["'\\' .. redis.call(\\'FLUSHALL\\') .. \\'", 'rediscallFLUSHALL'], ["' .. redis.call('FLUSHALL') .. '"]],
+        ];
+    }
+
+    public static function illegalValuesForWhere(): array
+    {
+        return [
+            [['id' => ["' .. redis.call('FLUSHALL') .. '" => 1]], ["'\\' .. redis.call(\\'FLUSHALL\\') .. \\'", 'rediscallFLUSHALL']],
+            [['id' => ['`id`=`id` and 1' => 1]], ["'`id`=`id` and 1'", 'ididand']],
+            [['id' => [
+                'legal' => 1,
+                '`id`=`id` and 1' => 1,
+            ]], ["'`id`=`id` and 1'", 'ididand']],
+            [['id' => [
+                'nested_illegal' => [
+                    'false or 1=' => 1
+                ]
+            ]], [], ['false or 1=']],
+        ];
+    }
+
+    public static function keyValueData(): array
+    {
+        return [
+            [123],
+            [-123],
+            [0],
+            ['test'],
+            ["test\r\ntest"],
+            [''],
+        ];
+    }
+
+    public static function multiSetExpiry(): array
+    {
+        return [[0], [2]];
+    }
+
+    public static function zRangeByScoreData(): array
+    {
+        return [
+            [
+                'members' => [
+                    ['foo', 1],
+                    ['bar', 2],
+                ],
+                'cases' => [
+                    // without both scores and limit
+                    ['0', '(1', null, null, null, null, []],
+                    ['1', '(2', null, null, null, null, ['foo']],
+                    ['2', '(3', null, null, null, null, ['bar']],
+                    ['(0', '2', null, null, null, null, ['foo', 'bar']],
+
+                    // with scores, but no limit
+                    ['0', '(1', 'WITHSCORES', null, null, null, []],
+                    ['1', '(2', 'WITHSCORES', null, null, null, ['foo', 1]],
+                    ['2', '(3', 'WITHSCORES', null, null, null, ['bar', 2]],
+                    ['(0', '2', 'WITHSCORES', null, null, null, ['foo', 1, 'bar', 2]],
+
+                    // with limit, but no scores
+                    ['0', '(1', null, 'LIMIT', 0, 1, []],
+                    ['1', '(2', null, 'LIMIT', 0, 1, ['foo']],
+                    ['2', '(3', null, 'LIMIT', 0, 1, ['bar']],
+                    ['(0', '2', null, 'LIMIT', 0, 1, ['foo']],
+
+                    // with both scores and limit
+                    ['0', '(1', 'WITHSCORES', 'LIMIT', 0, 1, []],
+                    ['1', '(2', 'WITHSCORES', 'LIMIT', 0, 1, ['foo', 1]],
+                    ['2', '(3', 'WITHSCORES', 'LIMIT', 0, 1, ['bar', 2]],
+                    ['(0', '2', 'WITHSCORES', 'LIMIT', 0, 1, ['foo', 1]],
+                ],
+            ],
+        ];
+    }
+}
